PLS-69 Refactor:Extract components
